requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
beautifulsoup4
requests
pytz
python-dateutil
pause
icalendar